pyproject.toml
[project]
name = "preferencia-interinos-programa"
version = "0.1.0"
description = ""
authors = [
    {name = "raistln",email = "samumarfon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit (>=1.45.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openai (>=1.79.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "geopy (>=2.4.1,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "preferencia_interinos_programa", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
